requirements.txt
python-dotenv==0.15.0
requests==2.25.1
selenium==3.141.0
